# public/docs/api-docs.json
{
  "info": {
    "title": "Playpost API",
    "description": "This is the documentation of the Playpost API",
    "version": "2.5.2",
    "termsOfService": "https://playpost.app/terms",
    "contact": {
      "name": "API Support",
      "url": "https://playpost.app/support",
      "email": "support@playpost.app"
    }
  },
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000/{basePath}",
      "description": "Development server",
      "variables": {
        "basePath": {
          "default": "v1"
        }
      }
    },
    {
      "url": "https://api.playpost.app/{basePath}",
      "description": "Production server",
      "variables": {
        "basePath": {
          "default": "v1"
        }
      }
    },
    {
      "url": "https://playpost-api-test.herokuapp.com/{basePath}",
      "description": "Test server",
      "variables": {
        "basePath": {
          "default": "v1"
        }
      }
    }
  ],
  "components": {
    "schemas": {
      "PatchOneUserRequestBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "api@playpost.app"
          },
          "currentPassword": {
            "type": "string",
            "example": "currentpassword!"
          },
          "newPassword": {
            "type": "string",
            "example": "supernewpassword!"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "authenticatedAt",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "4bee0e78-22e1-413d-9a65-8f7f95d9d0ca",
            "readOnly": true
          },
          "email": {
            "type": "string",
            "example": "api@playpost.app"
          },
          "password": {
            "type": "string",
            "example": "$2a$10$ETNlIhkfuKjTyl0VPdV5XOJQsAQBnLjaIIjasKFTPoC8lsQNPcja"
          },
          "resetPasswordToken": {
            "type": "string",
            "nullable": true,
            "example": "ABC123"
          },
          "activationToken": {
            "type": "string",
            "nullable": true,
            "example": "f3642b2891284604b4f7368d2831eff7"
          },
          "authenticatedAt": {
            "type": "string",
            "nullable": true,
            "example": "2020-02-14T03:30:10.000Z"
          },
          "requestResetPasswordAt": {
            "type": "string",
            "nullable": true,
            "example": "2020-02-14T03:30:10.000Z"
          },
          "resetPasswordAt": {
            "type": "string",
            "nullable": true,
            "example": "2020-02-14T03:30:10.000Z"
          },
          "activatedAt": {
            "type": "string",
            "nullable": true,
            "example": "2020-02-14T03:30:10.000Z"
          },
          "createdAt": {
            "type": "string",
            "example": "2020-02-14T03:30:10.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2020-02-14T03:30:10.000Z"
          }
        }
      },
...
    },
    "responses": {
...
      "BadRequestError": {
        "description": "A bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 400,
                  "description": "The HTTP Status code"
                },
                "message": {
                  "type": "string",
                  "description": "A string message summarizing the error"
                },
                "details": {
                  "type": "object",
                  "description": "An additional optional object containing more information about the error"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "A not found request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 404
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "UserResponse": {
        "description": "A User response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/User"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "ApiSecretAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Secret"
      }
    }
  },
...
  "paths": {
...
    "/user": {
      "get": {
        "operationId": "getOneUser",
        "tags": [
          "user"
        ],
        "summary": "Get the currently logged in user",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": [],
            "ApiSecretAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "patch": {
        "operationId": "patchOneUser",
        "tags": [
          "user"
        ],
        "summary": "Patch the currently logged in user",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": [],
            "ApiSecretAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchOneUserRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    }
  }
}